generate sitemap on the fly.
